Convert codegen to use Double instead of Float (#97)
Double is a wider precision type in Java and some AWS APIs expect Double, so it can't be cast to/from a Float, but the other way would work if we come across it.

+++ python/rpdk/java/pojo_resolver.py
@@ -82,7 +82,7 @@ class JavaPojoResolver:
             "string": "String",
             "integer": "Integer",
             "boolean": "Boolean",
-            "number": "Float",
+            "number": "Double",
         }
 
         return primitive_types_map[json_type]
